+++ CHANGELOG.md
@@ -5,8 +5,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+
+## [0.17.0-rc3]
+### Changed
+ - On Linux calculate CPU utilization in terms of logical, not physical, cores when possible.
+
+## [0.17.0-rc2]
+### Changed
+ - CPU percentage calculated in the same manner as Agent's
+
+## [0.17.0-rc1]
 ### Changed
 - Throttle metrics are now labeled with the respective generator's labels.
+- Observer now calculates CPU utilization with respect to target cgroup hard, soft limits. 
 
 ## [0.16.1]
 ### Changed